% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{add_l2nm}
\alias{add_l2nm}
\alias{add_l2sq}
\title{Add L2 Norm and Squared L2 Norm of Dimensions}
\usage{
add_l2nm(x)

add_l2sq(x)
}
\arguments{
\item{x}{A \code{tidy_fft} object containing frequency dimensions (\verb{.dim_*}) and associated metadata.}
}
\value{
A \code{tidy_fft} object with additional columns:
\itemize{
\item \strong{\code{l2nm}}: The L2 norm of the frequency dimensions.
\item \strong{\code{l2sq}}: The squared L2 norm of the frequency dimensions.
}
}
\description{
These functions compute and add L2 norms and squared L2 norms of the frequency dimensions
(\verb{.dim_*} columns) as additional columns to a \code{tidy_fft} object.
}
\details{
\itemize{
\item \strong{\code{add_l2nm()}}: Adds a column \code{l2nm} containing the L2 norm, calculated as the square root
of the sum of squares across \verb{.dim_*} columns.
\item \strong{\code{add_l2sq()}}: Adds a column \code{l2sq} containing the squared L2 norm, calculated as the sum
of squares across \verb{.dim_*} columns.
}
}
\examples{
matrix(1:9, 3) |>
  tidy_fft() |>
  print(n = 3) |>
  add_l2nm() |>
  print(n = 3) |>
  add_l2sq() |>
  print(n = 3)

}
\seealso{
\itemize{
\item \code{\link[=tidy_fft]{tidy_fft()}}
\item \code{\link[tibble:add_column]{tibble::add_column()}}
}
}
