% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{can_repr}
\alias{can_repr}
\title{Check if a \code{tidy_fft} object has a specific representation}
\usage{
can_repr(x, repr = c("polr", "rect", "cplx"))
}
\arguments{
\item{x}{A \code{tidy_fft} object.}

\item{repr}{The target representation to check (\code{"polr"}, \code{"rect"}, or \code{"cplx"}).}
}
\value{
A logical value (\code{TRUE} if the object has the specified representation, otherwise \code{FALSE}).
}
\description{
This function checks if the given \code{tidy_fft} object contains the specified
representation: polar (\code{"polr"}), rectangular (\code{"rect"}), or complex (\code{"cplx"}).
}
\examples{
tft <- tidy_fft(c(1, 0, -1, 0))
can_repr(tft, "cplx")  # Returns TRUE
can_repr(tft, "rect")  # Returns FALSE
}
