% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_repr}
\alias{change_repr}
\title{Change the representation of FFT results}
\usage{
change_repr(x, repr = c("polr", "rect", "cplx"), .keep = "unused")
}
\arguments{
\item{x}{A \code{tidy_fft} object.}

\item{repr}{The target representation (\code{"polr"}, \code{"rect"}, or \code{"cplx"}).}

\item{.keep}{Determines which columns to retain in the output.
Passed directly to \code{dplyr::mutate}. Options include:
\itemize{
\item \code{"all"}: Keep all columns (default).
\item \code{"used"}: Keep only columns used in the computation, plus the new ones.
\item \code{"unused"}: Keep only columns not used in the computation, plus the new ones.
\item \code{"none"}: Keep only the new columns.
}}
}
\value{
The modified object with the new representation.
}
\description{
Converts FFT results between different representations.
}
\examples{
fft_res <- tidy_fft(c(1, 0, -1, 0))
change_repr(fft_res, "rect", .keep = "none")
}
