% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross_spec}
\alias{cross_spec}
\title{Compute the Cross-Spectrum}
\usage{
cross_spec(a, b)
}
\arguments{
\item{a}{A \code{tidy_fft} object.}

\item{b}{A \code{tidy_fft} object with the same number of rows as \code{a}.}
}
\value{
A \code{tidy_fft} object with updated \code{fx} values representing the cross-spectrum.
}
\description{
This function computes the cross-spectrum between two \code{tidy_fft} objects by
multiplying the Fourier coefficients (\code{fx}) of the first object (\code{a}) with
the conjugate of the coefficients from the second object (\code{b}).
}
\details{
The cross-spectrum is calculated as:
\deqn{S_{ab}(k) = X_a(k) \cdot \overline{X_b(k)}}
where \eqn{X_a(k)} and \eqn{X_b(k)} are the Fourier coefficients of \code{a} and \code{b},
respectively, and \eqn{\overline{X_b(k)}} is the complex conjugate of \eqn{X_b(k)}.
}
\examples{
# Generate example tidy_fft objects
a <- tidy_fft(c(1, 0, -1, 0))
b <- tidy_fft(c(0, 1, 0, -1))

# Compute the cross-spectrum
cross_result <- cross_spec(a, b)
print(cross_result)

}
\seealso{
\code{\link[=filter_max_mod]{filter_max_mod()}} to filter the resulting cross-spectrum for the maximum magnitude.
}
