% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.as_fftab_obj}
\alias{.as_fftab_obj}
\title{Build a FFTAB Object}
\usage{
.as_fftab_obj(x, .is_normalized, .is_complex, ...)
}
\arguments{
\item{x}{The input object to convert, typically a tibble or data frame.}

\item{.is_normalized}{A logical value indicating if the input is normalized.}

\item{.is_complex}{A logical value indicating if the original data was complex.}

\item{...}{Additional attributes to include in the structured object, such as
metadata or specific attributes required for Fourier Transform analysis.}
}
\value{
The input object \code{x}, with the \code{fftab} class and any additional
attributes provided in \code{...}.
}
\description{
Converts an object into a \code{fftab} object with additional metadata attributes.
}
\keyword{internal}
