% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.as_tidy_fft_obj}
\alias{.as_tidy_fft_obj}
\title{Convert an Object to a Tidy FFT Object}
\usage{
.as_tidy_fft_obj(x, ...)
}
\arguments{
\item{x}{The input object to convert, typically a tibble or data frame.}

\item{...}{Additional attributes to include in the structured object, such
as metadata or specific attributes required for Fourier Transform analysis.}
}
\value{
The input object \code{x}, with the \code{tidy_fft} class and any additional
attributes provided in \code{...}.
}
\description{
This internal helper function applies the necessary structure and class
attributes to convert a given object into a \code{tidy_fft} object.
}
\details{
A \code{tidy_fft} object is a tibble-like structure with additional class
attributes used to represent Fourier Transform results in a tidy format.
}
\keyword{internal}
