% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.dim}
\alias{.dim}
\title{Retrieve the dimensions of an object}
\usage{
.dim(x)
}
\arguments{
\item{x}{An object to check for the \code{.dim} attribute.}
}
\value{
The value of the \code{.dim} attribute, or \code{NULL} if not present.
}
\description{
Retrieves the \code{.dim} attribute of the given object.
}
\keyword{internal}
