% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.fft}
\alias{.fft}
\title{Compute the Fast Fourier Transform (FFT) of a Vector}
\usage{
.fft(x, norm = FALSE)
}
\arguments{
\item{x}{A numeric vector representing the input signal to transform.}

\item{norm}{A logical value indicating whether to normalize the FFT output
by dividing it by the length of the input vector. Default is \code{FALSE}.}
}
\value{
A complex vector representing the FFT of the input signal.
}
\description{
This function computes the Fast Fourier Transform (FFT) of a numeric vector.
Optionally, it normalizes the result by dividing it by the length of the input vector.
}
\details{
The function wraps around the base R \code{stats::fft} function and provides an
option for normalization.
}
\examples{
x <- c(1, 2, 3, 4)
.fft(x)
.fft(x, norm = TRUE)

}
\keyword{internal}
