% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{.fourier_frequencies}
\alias{.fourier_frequencies}
\title{Compute Fourier frequencies for default inputs}
\usage{
.fourier_frequencies(x)
}
\arguments{
\item{x}{A scalar or vector representing the length of the sequence.}
}
\value{
A numeric vector containing the normalized Fourier frequencies.
}
\description{
Computes normalized Fourier frequencies for scalar or vector inputs, which
are evenly spaced between -0.5 and 0.5.
}
\keyword{internal}
