% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.frequency}
\alias{.frequency}
\title{Retrieve Frequency}
\usage{
.frequency(x)
}
\arguments{
\item{x}{A \code{tidy_fft} object or time series.}
}
\value{
The frequency value or \code{1} if not a time series.
}
\description{
Retrieves the frequency of a time series object.
}
\keyword{internal}
