% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_array}
\alias{.is_array}
\title{Check if an Object is an Array}
\usage{
.is_array(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
\code{TRUE} if the object has a \code{.dim} attribute, \code{FALSE} otherwise.
}
\description{
Determines if an object has a \code{.dim} attribute, indicating it is an array.
}
\keyword{internal}
