% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_complex}
\alias{.is_complex}
\title{Check if an object is complex}
\usage{
.is_complex(x)
}
\arguments{
\item{x}{An object to check for the \code{.is_complex} attribute.}
}
\value{
The value of the \code{.is_complex} attribute, or \code{NULL} if not present.
}
\description{
Retrieves the \code{.is_complex} attribute of the given object.
}
\keyword{internal}
