% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_normalized}
\alias{.is_normalized}
\title{Check if an Object is Normalized}
\usage{
.is_normalized(x)
}
\arguments{
\item{x}{An R object to check for normalization.}
}
\value{
A logical value indicating whether the \code{.is_normalized} attribute is \code{TRUE}.
Returns \code{NULL} if the attribute does not exist.
}
\description{
This internal helper function checks whether an object has been marked as normalized
by inspecting its \code{.is_normalized} attribute.
}
\details{
This function retrieves the value of the \code{.is_normalized} attribute from an object.
It is primarily used for internal validation purposes.
}
\keyword{internal}
