% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_ts}
\alias{.is_ts}
\title{Check if an object is a time series}
\usage{
.is_ts(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
\code{TRUE} if the object has a \code{.tsp} attribute, \code{FALSE} otherwise.
}
\description{
Determines if the object has a \code{.tsp} attribute, indicating it is a time series.
}
\keyword{internal}
