% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.nyquist}
\alias{.nyquist}
\title{Compute Nyquist Frequency}
\usage{
.nyquist(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
The Nyquist frequency (\code{.frequency(x) / 2}).
}
\description{
Computes the Nyquist frequency for an object.
}
\keyword{internal}
