% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.shift_phase}
\alias{.shift_phase}
\title{Shift Phase}
\usage{
.shift_phase(x, shift)
}
\arguments{
\item{x}{An R object.}

\item{shift}{A numeric value indicating the phase shift.}
}
\description{
Adjusts the phase of Fourier coefficients.
}
\keyword{internal}
