% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.size}
\alias{.size}
\title{Retrieve Object Size}
\usage{
.size(x)
}
\arguments{
\item{x}{An R object to check for size.}
}
\value{
The \code{.size} attribute or \code{NULL} if not present.
}
\description{
Retrieves the \code{.size} attribute from an object.
}
\keyword{internal}
