% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.tsp}
\alias{.tsp}
\title{Retrieve the time series parameters of an object}
\usage{
.tsp(x)
}
\arguments{
\item{x}{An object to check for the \code{.tsp} attribute.}
}
\value{
The value of the \code{.tsp} attribute, or \code{NULL} if not present.
}
\description{
Retrieves the \code{.tsp} attribute of the given object.
}
\keyword{internal}
