% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.variance}
\alias{.variance}
\title{Compute Variance}
\usage{
.variance(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A numeric variance value.
}
\description{
Computes the variance of Fourier coefficients.
}
\keyword{internal}
