% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftab-package.R
\docType{package}
\name{fftab-package}
\alias{fftab-package}
\title{fftab: Tidy Manipulation of Fourier Transformed Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'fftab' package stores Fourier coefficients in a tibble and allows their manipulation in various ways. Functions are available for converting between complex, rectangular ('re', 'im'), and polar ('mod', 'arg') representations, as well as for extracting components as vectors or matrices. Inputs can include vectors, time series, and arrays of arbitrary dimensions, which are restored to their original form when inverting the transform. Since 'fftab' stores Fourier frequencies as columns in the tibble, many standard operations on spectral data can be easily performed using tidy packages like 'dplyr'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thk686/fftab}
  \item \url{https://thk686.github.io/fftab/}
  \item Report bugs at \url{https://github.com/thk686/fftab/issues}
}

}
\author{
\strong{Maintainer}: Timothy Keitt \email{tkeitt@gmail.com}

}
\keyword{internal}
