% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fourier_frequencies.default}
\alias{fourier_frequencies.default}
\title{Compute Fourier frequencies for default inputs}
\usage{
\method{fourier_frequencies}{default}(x)
}
\arguments{
\item{x}{A scalar or vector representing the length of the sequence.}
}
\value{
A numeric vector of normalized Fourier frequencies.
}
\description{
Computes normalized Fourier frequencies for scalar or vector inputs.
}
\examples{
fourier_frequencies(8)
}
