% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fourier_frequencies}
\alias{fourier_frequencies}
\title{Compute Fourier frequencies}
\usage{
fourier_frequencies(x)
}
\arguments{
\item{x}{The input (scalar, vector, or time series object).}
}
\value{
A vector of Fourier frequencies, normalized or scaled based on the input type.
}
\description{
This is a generic function for computing Fourier frequencies.
}
