% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rect}
\alias{get_rect}
\alias{get_polr}
\title{Extract Rectangular or Polar Components}
\usage{
get_rect(x)

get_polr(x)
}
\arguments{
\item{x}{A \code{tidy_fft} object containing FFT results.}
}
\value{
\itemize{
\item \strong{\code{get_rect}}: A matrix with two columns:
\itemize{
\item \code{re}: The real part of the coefficients.
\item \code{im}: The imaginary part of the coefficients.
}
\item \strong{\code{get_polr}}: A matrix with two columns:
\itemize{
\item \code{mod}: The magnitude of the coefficients.
\item \code{arg}: The phase angle of the coefficients, in radians.
}
}
}
\description{
The \code{get_rect} and \code{get_polr} functions extract specific components from a
\code{tidy_fft} object, representing the Fourier coefficients in either rectangular
or polar form.
}
\examples{
fft_result <- tidy_fft(c(1, 0, -1, 0))

# Extract rectangular components
rect_values <- get_rect(fft_result)
print(rect_values)

# Extract polar components
polr_values <- get_polr(fft_result)
print(polr_values)

}
\seealso{
\code{\link[=get_fx]{get_fx()}}, \code{\link[=get_re]{get_re()}}, \code{\link[=get_mod]{get_mod()}}, \code{\link[=to_rect]{to_rect()}}, \code{\link[=to_polr]{to_polr()}}
}
