% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_cplx}
\alias{has_cplx}
\title{Check if a \code{tidy_fft} object has a specific representation}
\usage{
has_cplx(x)
}
\arguments{
\item{x}{A \code{tidy_fft} object.}

\item{repr}{The target representation to check (\code{'polr'}, \code{'rect'}, or \code{'cplx'}).}
}
\value{
A logical value (\code{TRUE} if the object has the specified representation, otherwise \code{FALSE}).

A vector of possible representations
}
\description{
This function checks if the given \code{tidy_fft} object contains the specified
representation: polar (\code{'polr'}), rectangular (\code{'rect'}), or complex (\code{'cplx'}).
}
\examples{
tft <- tidy_fft(c(1, 0, -1, 0))
can_repr(tft, "cplx") # Returns TRUE
can_repr(tft, "rect") # Returns FALSE
}
