% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fftab}
\alias{plot.fftab}
\title{Plot the modulus of FFT results}
\usage{
\method{plot}{fftab}(x, ...)
}
\arguments{
\item{x}{A \code{fftab} object. This should contain Fourier-transformed data.}

\item{...}{Additional arguments passed to \code{ggplot2::ggplot}.}
}
\value{
A \code{ggplot} object representing the modulus of FFT results plotted against the frequencies.
The plot shows the modulus (\code{mod}) on the y-axis and frequency values on the x-axis.
}
\description{
Plots the modulus of the FFT results against the frequencies.
}
