% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.tidy_fft}
\alias{plot.tidy_fft}
\title{Plot the modulus of FFT results}
\usage{
\method{plot}{tidy_fft}(x, ...)
}
\arguments{
\item{x}{A \code{tidy_fft} object.}

\item{...}{passed to ggplot.}
}
\description{
Plots the modulus of the FFT results against the frequencies.
}
