% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{repr}
\alias{repr}
\title{Retrieve the current representation of a tidy_fft object}
\usage{
repr(x)
}
\arguments{
\item{x}{A \code{tidy_fft} object.}
}
\value{
The current representation (\code{"polr"}, \code{"rect"}, or \code{"cplx"}).
}
\description{
Retrieve the current representation of a tidy_fft object
}
