% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{tidy_fft.default}
\alias{tidy_fft.default}
\title{Perform FFT for default inputs}
\usage{
\method{tidy_fft}{default}(x, repr = c("polr", "rect", "cplx"))
}
\arguments{
\item{x}{A numeric vector.}

\item{repr}{The desired representation of the FFT result.}
}
\value{
A tibble containing Fourier frequencies and FFT values in the specified format.
}
\description{
Computes the FFT for a numeric vector and returns a tidy result.
}
\examples{
tidy_fft(c(1, 0, -1, 0), repr = "rect")
}
