% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{tidy_fft.ts}
\alias{tidy_fft.ts}
\title{Perform FFT for time series inputs}
\usage{
\method{tidy_fft}{ts}(x, repr = c("polr", "rect", "cplx"))
}
\arguments{
\item{x}{A time series (\code{ts}) object.}

\item{repr}{The desired representation of the FFT result.}
}
\value{
A tibble containing Fourier frequencies and FFT values in the specified format.
}
\description{
Computes the FFT for a time series object and returns a tidy result.
}
\examples{
ts_obj <- ts(sin(1:10), frequency = 12)
tidy_fft(ts_obj, repr = "polr")
}
