% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{tidy_ifft}
\alias{tidy_ifft}
\title{Perform inverse FFT on a tidy result}
\usage{
tidy_ifft(x)
}
\arguments{
\item{x}{A \code{tidy_fft} object containing FFT results.}
}
\value{
A vector or time series object reconstructed from the FFT results.
}
\description{
This is a generic function for performing the inverse FFT on a \code{tidy_fft} object.
}
