% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{tidy_ifft.tidy_fft}
\alias{tidy_ifft.tidy_fft}
\title{Perform inverse FFT for default inputs}
\usage{
\method{tidy_ifft}{tidy_fft}(x)
}
\arguments{
\item{x}{A \code{tidy_fft} object.}
}
\value{
The reconstructed signal as a vector or time series object.
}
\description{
Performs the inverse FFT and reconstructs the original signal.
}
\examples{
fft_res <- tidy_fft(c(1, 0, -1, 0), repr = "cplx")
tidy_ifft(fft_res)
}
