% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr.R
\name{to_angf}
\alias{to_angf}
\alias{to_cycf}
\title{Convert between cyclic and angular frequencies}
\usage{
to_angf(x)

to_cycf(x)
}
\arguments{
\item{x}{An \code{fftab} object containing frequency dimensions. Must include
columns prefixed with \code{.dim_} and have an attribute \code{.is_angular} indicating
the frequency type.}
}
\value{
An \code{fftab} object with dimensions scaled appropriately and the \code{.is_angular}
attribute updated.
}
\description{
These functions convert dimensions of frequency from cyclic (measured in cycles)
to angular (measured in radians) or vice versa. This transformation scales
dimensions by a factor of 2 * pi.
\itemize{
\item \code{to_angf()}: Converts from cyclic to angular frequency.
\item \code{to_cycf()}: Converts from angular to cyclic frequency.
}
}
\examples{
# Convert cyclic to angular frequencies
rnorm(64) |>
  fftab() |>
  to_angf() |>
  to_rect() |>
  dplyr::slice_max(abs(.dim_1), n = 5)

# Convert angular back to cyclic frequencies
rnorm(64) |>
  fftab() |>
  to_angf() |>
  to_cycf() |>
  to_rect() |>
  dplyr::slice_max(abs(.dim_1), n = 5)

}
