% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{change_repr}
\alias{change_repr}
\title{Change the representation of FFT results}
\usage{
change_repr(x, repr = c("polr", "rect", "cplx"))
}
\arguments{
\item{x}{A \code{tidy_fft} object.}

\item{repr}{The target representation (\code{"polr"}, \code{"rect"}, or \code{"cplx"}).}
}
\value{
The modified object with the new representation.
}
\description{
Converts FFT results between different representations.
}
\examples{
fft_res <- tidy_fft(c(1, 0, -1, 0), repr = "cplx")
change_repr(fft_res, "rect")
}
