% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.as_tidy_fft_obj}
\alias{.as_tidy_fft_obj}
\title{Build a Tidy FFT Object}
\usage{
.as_tidy_fft_obj(x, .is_normalized, .is_complex, ...)
}
\arguments{
\item{x}{The input object to convert, typically a tibble or data frame.}

\item{.is_normalized}{A logical value indicating if the input is normalized.}

\item{.is_complex}{A logical value indicating if the original data was complex.}

\item{...}{Additional attributes to include in the structured object, such as
metadata or specific attributes required for Fourier Transform analysis.}
}
\value{
The input object \code{x}, with the \code{tidy_fft} class and any additional
attributes provided in \code{...}.
}
\description{
Converts an object into a \code{tidy_fft} object with additional metadata attributes.
}
\seealso{
\code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\keyword{internal}
