% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.dim}
\alias{.dim}
\title{Retrieve Object Dimensions}
\usage{
.dim(x)
}
\arguments{
\item{x}{An R object to check for dimensions.}
}
\value{
The \code{.dim} attribute or \code{NULL} if not present.
}
\description{
Retrieves the \code{.dim} attribute from an object.
}
\keyword{internal}
