% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.is_complex}
\alias{.is_complex}
\title{Check if an Object is Complex}
\usage{
.is_complex(x)
}
\arguments{
\item{x}{An R object to check for complexity.}
}
\value{
A logical value indicating whether the \code{.is_complex} attribute is \code{TRUE}.
}
\description{
Checks whether an object has the \code{.is_complex} attribute.
}
\keyword{internal}
