% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.is_normalized}
\alias{.is_normalized}
\title{Check if an Object is Normalized}
\usage{
.is_normalized(x)
}
\arguments{
\item{x}{An R object to check for normalization.}
}
\value{
A logical value indicating whether the \code{.is_normalized} attribute is \code{TRUE}.
Returns \code{NULL} if the attribute does not exist.
}
\description{
Checks whether an object has the \code{.is_normalized} attribute.
}
\keyword{internal}
