% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_ts}
\alias{.is_ts}
\title{Check if an Object is a Time Series}
\usage{
.is_ts(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
\code{TRUE} if the object has a \code{.tsp} attribute, \code{FALSE} otherwise.
}
\description{
Determines if an object has a \code{.tsp} attribute.
}
\keyword{internal}
