% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.size}
\alias{.size}
\title{Retrieve the size of an object}
\usage{
.size(x)
}
\arguments{
\item{x}{An object to check for the \code{.size} attribute.}
}
\value{
The value of the \code{.size} attribute, or \code{NULL} if not present.
}
\description{
Retrieves the \code{.size} attribute of the given object.
}
\keyword{internal}
