% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.tsp}
\alias{.tsp}
\title{Retrieve Time Series Parameters}
\usage{
.tsp(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
The \code{.tsp} attribute or \code{NULL} if not present.
}
\description{
Retrieves the \code{.tsp} attribute from an object.
}
\keyword{internal}
