% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftab-package.R
\docType{package}
\name{fftab-package}
\alias{fftab-package}
\title{fftab: Tidy Manipulation of Fourier Transformed Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functions for storing fourier coefficients in a tibble and manipulating them in various ways. Functions are provided for changing between complex, rectangular (re, im), and polar (mod, arg) representations are provided, as well as functions for extracting components as vectors or matrices. Inputs can be vectors, time series, and arrays of arbitrary dimension, and these are restored to their original form when inverting the transform. Because fftab stores the fourier frequencies as columns in the tibble, many standard operations on spectral data can be easily computed using tidy packages, including dplyr.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thk686/fftab}
  \item \url{https://thk686.github.io/fftab/}
  \item Report bugs at \url{https://github.com/thk686/fftab/issues}
}

}
\author{
\strong{Maintainer}: Timothy Keitt \email{tkeitt@gmail.com}

}
\keyword{internal}
