% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftab-package.R
\docType{package}
\name{fftab-package}
\alias{fftab-package}
\title{fftab: Tidy manipulation of fourier transformed data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functions for storing fourier coefficients in a tibble and manipulating them in various ways. The main different between tidy_fft and base R is that tidy_fft computes the normalized fourier frequencies in cycles per stride, i.e., cycles per interval between samples, which is often a point of confusion for new users and convenient for plotting, etc. Because the frequencies and coefficients are stored in a tibble (data.frame), easy processing with, e.g., dplyr is facilitated.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thk686/fftab}
  \item \url{https://thk686.github.io/fftab/}
  \item Report bugs at \url{https://github.com/thk686/fftab/issues}
}

}
\author{
\strong{Maintainer}: Timothy Keitt \email{tkeitt@gmail.com}

}
\keyword{internal}
