% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{fourier_frequencies}
\alias{fourier_frequencies}
\title{Compute Fourier frequencies}
\usage{
fourier_frequencies(x)
}
\arguments{
\item{x}{The input object (scalar, vector, matrix, time series, or array).}
}
\value{
A numeric vector for 1D inputs or a tibble for multidimensional inputs,
containing the Fourier frequencies normalized or scaled based on the input type.
}
\description{
This is a generic function for computing Fourier frequencies for various
types of inputs, such as scalars, vectors, matrices, time series, or arrays.
}
