% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_repr}
\alias{get_repr}
\title{Retrieve the current representation of a tidy_fft object}
\usage{
get_repr(x)
}
\arguments{
\item{x}{A \code{tidy_fft} object.}
}
\value{
A vector of possible representations
}
\description{
Retrieve the current representation of a tidy_fft object
}
