% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fftab}
\alias{plot.fftab}
\title{Plot the modulus of FFT results}
\usage{
\method{plot}{fftab}(x, ...)
}
\arguments{
\item{x}{A \code{fftab} object.}

\item{...}{passed to ggplot.}
}
\description{
Plots the modulus of the FFT results against the frequencies.
}
