% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{remove_dc}
\alias{remove_dc}
\alias{remove_symmetric}
\alias{split_symmetric}
\title{Remove DC Component and Symmetric Frequencies}
\usage{
remove_dc(x)

remove_symmetric(x)

split_symmetric(x)
}
\arguments{
\item{x}{A \code{tidy_fft} object containing Fourier coefficients and associated metadata.}
}
\value{
A \code{tidy_fft} object with filtered coefficients.
}
\description{
These functions operate on \code{tidy_fft} objects to manipulate and filter Fourier coefficients.
}
\details{
\itemize{
\item \code{remove_dc()}: Removes the DC (zero-frequency) component.
\item \code{remove_symmetric()}: Removes symmetric, complex-conjugate frequencies.
\item \strong{\code{remove_dc()}}: Filters out rows where all \verb{.dim_*} columns have a value of \code{0}.
\item \strong{\code{remove_symmetric()}}:
\itemize{
\item For real-valued signals, it filters out redundant, complex-conjugate frequencies
\item For complex-valued signals, no filtering is applied as symmetry isn't relevant.
}
}
}
\examples{
fft_x <- tidy_fft(sin(seq(0, 2 * pi, length.out = 128)))
fft_no_dc <- remove_dc(fft_x)
fft_no_sym <- remove_symmetric(fft_x)

}
\seealso{
\itemize{
\item \code{\link[dplyr:filter]{dplyr::filter()}}
\item \code{\link[=tidy_fft]{tidy_fft()}}
}
}
