% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{tidy_fft.default}
\alias{tidy_fft.default}
\title{Perform FFT for default inputs}
\usage{
\method{tidy_fft}{default}(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A tibble containing Fourier frequencies and FFT values in the specified format.
}
\description{
Computes the FFT for a numeric vector and returns a tidy result.
}
\examples{
tidy_fft(c(1, 0, -1, 0))
}
