% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{tidy_fft}
\alias{tidy_fft}
\title{Perform FFT and return a tidy result}
\usage{
tidy_fft(x)
}
\arguments{
\item{x}{Input object (vector, time series, or array).}
}
\value{
A tibble containing Fourier frequencies and FFT values in the specified format.
}
\description{
This is a generic function to compute the FFT and return the result in a tidy format.
}
