% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{tidy_fft}
\alias{tidy_fft}
\title{Perform FFT and return a tidy result}
\usage{
tidy_fft(x, repr = c("polr", "rect", "cplx"))
}
\arguments{
\item{x}{Input object (vector or \code{ts} object).}

\item{repr}{The desired representation of the FFT result: \code{"polr"}, \code{"rect"}, or \code{"cplx"}.}
}
\value{
A tibble containing Fourier frequencies and FFT values in the specified format.
}
\description{
This is a generic function to compute the FFT and return the result in a tidy format.
}
