% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyfft-package.R
\docType{package}
\name{tidyfft-package}
\alias{tidyfft}
\alias{tidyfft-package}
\title{tidyfft: Tidy manipulation of fourier transformed data}
\description{
Provides functions for storing fourier coefficients in a tibble and manipulating them in various ways.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thk686/tidyfft}
  \item Report bugs at \url{https://github.com/thk686/tidyfft/issues}
}

}
\author{
\strong{Maintainer}: Timothy Keitt \email{tkeitt@gmail.com}

}
\keyword{internal}
