% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_cplx}
\alias{to_cplx}
\alias{to_rect}
\alias{to_polr}
\title{Convert a \code{tidy_fft} Object Between Representations}
\usage{
to_cplx(x, .keep = "unused")

to_rect(x, .keep = "unused")

to_polr(x, .keep = "unused")
}
\arguments{
\item{x}{A \code{tidy_fft} object.}

\item{.keep}{Specifies which columns to retain. See \code{\link[=mutate]{mutate()}}.}
}
\value{
A modified \code{tidy_fft} object containing the specified representation:
\itemize{
\item \strong{\code{to_cplx()}}: Adds the \code{fx} column for complex values.
\item \strong{\code{to_rect()}}: Adds the \code{re} and \code{im} columns for rectangular components.
\item \strong{\code{to_polr()}}: Adds the \code{mod} and \code{arg} columns for polar components.
}
}
\description{
These functions convert a \code{tidy_fft} object to a specified representation:
\itemize{
\item \strong{\code{to_cplx()}}: Converts to complex representation (\code{fx}).
\item \strong{\code{to_rect()}}: Converts to rectangular representation (\code{re}, \code{im}).
\item \strong{\code{to_polr()}}: Converts to polar representation (\code{mod}, \code{arg}).
}
}
\details{
\itemize{
\item \strong{\code{to_cplx()}}: Converts from rectangular (\code{re}, \code{im}) or polar (\code{mod}, \code{arg}) components to complex form.
\item \strong{\code{to_rect()}}: Converts from complex (\code{fx}) or polar components to rectangular form.
\item \strong{\code{to_polr()}}: Converts from complex (\code{fx}) or rectangular components to polar form.
}
}
\examples{
tft <- tidy_fft(c(1, 0, -1, 0))

# Convert to different representations
tft_cplx <- to_cplx(tft) # Complex representation
tft_rect <- to_rect(tft_cplx) # Rectangular representation
tft_polr <- to_polr(tft_cplx) # Polar representation

# Print results
print(tft_cplx)
print(tft_rect)
print(tft_polr)

}
\seealso{
\code{\link[=can_repr]{can_repr()}}, \code{\link[=get_repr]{get_repr()}}
}
